[#1515] Moved filter logic to filters.py
