+++ akvo/rsr/filters.py
@@ -10,14 +10,12 @@ see < http://www.gnu.org/licenses/agpl.html >.
 import django_filters
 
 from copy import deepcopy
-
-from .models import Project, Organisation, Category, ProjectUpdate
-from .m49 import M49_CODES, M49_HIERARCHY
-
+from django.utils.translation import ugettext_lazy as _
 from akvo.codelists.store.codelists_v201 import SECTOR_CATEGORY
 from akvo.utils import codelist_choices
-
-from django.utils.translation import ugettext_lazy as _
+from .models import (Category, Organisation, OrganisationLocation, Project,
+                     ProjectLocation, ProjectUpdate, ProjectUpdateLocation)
+from .m49 import M49_CODES, M49_HIERARCHY
 
 ANY_CHOICE = (('', _('All')), )
 
@@ -31,7 +29,8 @@ def sectors():
 
 
 def get_orgs():
-    orgs = list(Organisation.objects.all().values_list('id', 'name', flat=False))
+    orgs = list(Organisation.objects.all().values_list(
+                'id', 'name', flat=False))
     return [('', _('All'))] + orgs
 
 
@@ -66,6 +65,58 @@ def filter_m49(queryset, value):
     return queryset.filter(primary_location__country__iso_code__in=countries)
 
 
+def get_id_for_iso(i):
+    """From an iso_code e.g. 'SE' get the identifier."""
+    i = [k for k, v in M49_HIERARCHY.iteritems() if i in v]
+    if not i:
+        return None
+    else:
+        return i.pop()
+
+
+def get_locations(location, locations):
+    """Based on one location (country or group as Europe) get all the"""
+    l = get_id_for_iso(location)
+    if isinstance(l, basestring):
+        return locations
+    elif l is 1:
+        return locations
+    elif l is None:
+        return locations
+    else:
+        locations.append(l)
+        return get_locations(l, locations)
+
+
+def location_choices(qs):
+    """From a queryset get possible location filter choices"""
+
+    if qs.model is Project:
+        location_model = ProjectLocation
+    elif qs.model is ProjectUpdate:
+        location_model = ProjectUpdateLocation
+    elif qs.model is Organisation:
+        location_model = OrganisationLocation
+
+    locations_qs = location_model.objects.filter(
+        location_target__in=qs).order_by('country__id').distinct('country__id')
+
+    locations = []
+    for location in locations_qs:
+        country = get_id_for_iso(location.country.iso_code.upper())
+        locations.append(country)
+        locations.extend(get_locations(country, []))
+
+    choices = [tup for tup in M49_CODES if any(
+        unicode(i) in tup for i in locations)]
+    return [M49_CODES[0]] + choices  # Add the world to the choices
+
+
+def build_choices(qs):
+    """Build choices from queryset and add an All option"""
+    return [('', _('All'))] + list(qs.values_list('id', 'name', flat=False))
+
+
 class ProjectFilter(django_filters.FilterSet):
 
     category = django_filters.ChoiceFilter(
@@ -97,7 +148,8 @@ class ProjectFilter(django_filters.FilterSet):
         name='title')
 
     organisation = django_filters.ChoiceFilter(
-        choices=get_orgs(),
+        # choices=get_orgs(),
+        choices=(('', 'None')),
         label=_(u'organisation'),
         name='partners__id')
